Refactor code and import image asset
This commit involves refactoring some parts of the App module and importing an image asset. From now on, the 'React' import is replaced with only 'useState' since 'React' is not in use. An image asset, 'bird.png', is also imported at the top of the file and correctly referenced in the image source attribute.